✨ [Feature] 수량 변화에 따른 가격 변동 기능 추가
